39 fix after review (#40)
* [39] use same way for setting integer db type and primary db key

* [39] use slf4j annotation to inject logger

* [39] remove validation from timestamp because it is set above

* [39] remove unused dependency

* [39] remove id column from many to many films_to_directors table

+++ src/main/java/ru/yandex/practicum/filmorate/dao/ReviewDbStorage.java
@@ -8,7 +8,7 @@ public interface ReviewDbStorage {
 
     List<Review> getFilmsReviews(int count);
 
-    Review getReview(long id);
+    Review getReview(int id);
 
     void deleteReview(int id);
 
@@ -26,5 +26,5 @@ public interface ReviewDbStorage {
 
     List<Review> getFilmsReviewsById(int filmId, int count);
 
-    List<Long> getListId();
+    List<Integer> getListId();
 }
